Edited some of the Thruster code to make it better

+++ core/src/rami/project/grey/core/entity/consumable/IAttachable.java
@@ -3,5 +3,5 @@ package rami.project.grey.core.entity.consumable;
 import rami.project.grey.core.entity.IEntity;
 
 public interface IAttachable extends IEntity {
-    byte getHealthPenalty();
+    short getHealthPenalty();
 }
